// client/app/shared/models/recipe.model.ts
export class Recipe {
    _id?: string;
    title?: String;
    description?: String;
    service_quantity?: Number;
    serving_units?: String;
    preparation_time?: String;
    cuisine?: String;
    main_ingredients?: String;
    type_of_recipe?: String;
    meal?: String;
    preparation_method?: String;
    dietary_consideration?: String;
    video_type?: String;
    youtube_link?: String;
    youtube_video_id?: String;
    manuel_rating?: Number;
    youtuble_embed_url: String;
    youtube_rating?: Number;
    youtube_total_comments?: Number;
    youtube_publisher_name?: String;
    youtube_publish_date?: Date;
    youtube_original_date?: Text;
    created_ip?: String;
    last_update_ip?: String;
    holiday?: String;
    account_id?: String;
}